requirements.txt
tensorflow==2.0.1
#h5py==2.9.0
#keras_preprocessing==1.0.5
keras_applications==1.0.8
#gast==0.2.2
futures
protobuf
pybind11
flask
pillow
gunicorn
